[chores] Improved solution of OpenVPN parsing bug
node ID will be always just the host unless
there are multiple hosts with the same address,
in which case host:port will be used.

This makes the network graph cleaner because ports
can change often in OpenVPN and as a result the graph
will show new nodes when in the truth is always
the same node on a different port.